requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
python-dotenv==0.19.2
python-telegram-bot==13.10
pytube==11.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
